Basic idea of fetching messages
Now we have two users who play secret number with each other but they
are incredibly good at it and get the number right after some tries. It
is also stated that something is making a request. In a real case
scenario, a request would be made every 5 seconds or so and messages
would be fetched from the server. As of now, the initial list of
messages is empty and the client has to fetch the initial messages from
the server. This should be improved in the future